# packages/server/swagger/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Letter Forge Engine API",
    "description": "API Documentation for Letter Forge Engine",
    "version": "1.0.0"
  },
  "host": "localhost:8888",
  "basePath": "/letter-forge/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT bearer token (format: \"Bearer <token>\")"
    }
  },
...
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register a new account",
        "description": "Creates an account and returns a JWT token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "nickname": {
                  "type": "string"
                },
                "password": {
                  "type": "string",
                  "minLength": 8
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log in and receive a token",
        "description": "Authenticates an account and returns a JWT token",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": [
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Invalid email or password"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Get the authenticated user",
        "description": "Returns the currently authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/games": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Create a new game",
        "description": "Creates a new game with specified mode (normal_mode, time_attack, survival_mode, or chain_mode)",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "mode": {
                  "type": "string",
                  "enum": [
                    "normal_mode",
                    "time_attack",
                    "survival_mode",
                    "chain_mode"
                  ],
                  "description": "Game mode"
                },
                "letterCount": {
                  "type": "integer",
                  "description": "Number of letters (2 to 5)"
                },
                "letters": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "description": "Predetermined letters (optional)"
                },
                "userId": {
                  "type": "string",
                  "description": "User ID (optional)"
                }
              },
              "required": [
                "mode"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/games/{gameId}": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Load a game",
        "description": "Loads an existing game by its ID",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "games"
        ],
        "summary": "Delete a game",
        "description": "Permanently deletes an in-progress game. Completed games cannot be deleted.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Game deleted"
          },
          "400": {
            "description": "Completed games cannot be deleted"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Not the game owner"
          },
          "404": {
            "description": "Game not found"
          }
        }
      }
    },
    "/games/{gameId}/submit": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Submit a word",
        "description": "Submits a word for a specific game",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "word": {
                  "type": "string",
                  "example": "sea"
                }
              },
              "required": [
                "word"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/games/{gameId}/reset": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Reset letters",
        "description": "Generates new random letters and clears used words",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "letterCount": {
                  "type": "integer",
                  "example": 3
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/games/{gameId}/complete": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Complete the game",
        "description": "Marks the game as completed",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/games/{gameId}/batch": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Refill letter batch",
        "description": "Generates a new batch of 10 letter sets for batch-enabled modes when current batch is nearly exhausted",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Game mode does not support batch generation"
          }
        }
      }
    },
    "/games/{gameId}/result": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Get game result",
        "description": "Returns the final result of the game",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/games/{gameId}/leaderboard": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Submit game score to leaderboard",
        "description": "Submits the final game score to the leaderboard",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "period": {
                  "type": "string",
                  "enum": [
                    "daily",
                    "weekly",
                    "all_time"
                  ],
                  "default": "all_time"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}